refactor(shared/components/modal): 컴포넌트 관련 파일 kebab case로 변경, index.ts 추가
