+++ packages/shared/components/src/Modal/index.ts
@@ -1,0 +1,9 @@
+/* eslint-disable @stylistic/padding-line-between-statements, react-refresh/only-export-components */
+export { Modal } from './modal';
+export { ModalBody, type ModalBodyProps } from './modal-body';
+export { ModalContent, type ModalContentProps } from './modal-content';
+export { ModalFooter, type ModalFooterProps } from './modal-footer';
+export { ModalHeader, type ModalHeaderProps } from './modal-header';
+export { ModalOverlay, type ModalOverlayProps } from './modal-overlay';
+export { ModalTitle, type ModalTitleProps } from './modal-title';
+export { ModalTopbar, type ModalTopbarProps } from './modal-topbar';
