+++ packages/shared/components/src/Modal/modal-body.tsx
@@ -1,0 +1,21 @@
+import { type HTMLFavolinkProps, favolink, forwardRef } from '@favolink/system';
+import { classNames } from '@favolink/utils';
+
+export type ModalBodyProps = HTMLFavolinkProps<'main'>;
+
+export const ModalBody = forwardRef<ModalBodyProps, 'main'>(
+  function ModalBody(props, ref) {
+    const { children, className, ...restProps } = props;
+
+    return (
+      <favolink.main
+        as="main"
+        {...restProps}
+        ref={ref}
+        className={classNames('favolink-modal__body', className)}
+      >
+        {children}
+      </favolink.main>
+    );
+  },
+);
