+++ packages/shared/components/src/Modal/modal-topbar.tsx
@@ -1,0 +1,34 @@
+import { type HTMLFavolinkProps, favolink, forwardRef } from '@favolink/system';
+import { classNames } from '@favolink/utils';
+import * as styles from './modal-topbar.styles.css';
+import { useModalContext } from './modal.context';
+import { Text } from '../text';
+
+export type ModalTopbarProps = HTMLFavolinkProps<'div'> & {
+  layout: styles.ModalTopbarLayout;
+};
+
+export const ModalTopbar = forwardRef<ModalTopbarProps, 'div'>(
+  function ModalTopBar(props, ref) {
+    const { children, layout, className, ...restProps } = props;
+
+    const { onClose } = useModalContext();
+
+    const isCouple = layout === 'couple';
+
+    return (
+      <favolink.div
+        {...restProps}
+        ref={ref}
+        className={classNames(
+          `favolink-modal__topbar`,
+          styles.modalTopbarLayout[layout],
+          className,
+        )}
+      >
+        {isCouple && <Text>{children}</Text>}
+        <Text onClick={onClose}>닫기</Text>
+      </favolink.div>
+    );
+  },
+);
