+++ packages/shared/components/src/Modal/modal.context.ts
@@ -1,0 +1,12 @@
+import { createContext } from '@favolink/system';
+
+export type ModalContextDefaultValue = {
+  onClose: () => void;
+  closeOnOverlayClick?: boolean;
+};
+
+export const [ModalContextProvider, useModalContext] =
+  createContext<ModalContextDefaultValue>({
+    onClose: () => {},
+    closeOnOverlayClick: false,
+  });
